refactor: move utils pkg to internal
