+++ internal/utils/mocks/Encryptor.go
