+++ internal/utils/pass_encryptor.go
